backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
drf-spectacular==0.27.2
Pillow==11.1.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
